+++ tests/Authorization/AuthorizationIntegrationTest.php
@@ -36,11 +36,49 @@ class AuthorizationIntegrationTest extends AbstractTestCase
 
     public function testSubmissionLevelAuthorizationAddAndRemoveSubmissionResource()
     {
-        $form = $this->addFormWithSubmissionLevelAuthorization();
+        $form = $this->testEntityManager->addForm(
+            grantBasedSubmissionAuthorization: true,
+            allowedSubmissionStates: Submission::SUBMISSION_STATE_SUBMITTED,
+            actionsAllowedWhenSubmitted: [AuthorizationService::READ_SUBMISSION_ACTION],
+        );
 
+        // for submitted submissions, the creator should have only grants for the form's actionsAllowedWhenSubmitted.
+        // test entering the submitted state directly during add (POST).
         $submission = new Submission();
         $submission->setDataFeedElement('{"givenName": "John", "familyName": "Doe"}');
         $submission->setForm($form);
+        $submission->setSubmissionState(Submission::SUBMISSION_STATE_SUBMITTED);
+
+        $submission = $this->formalizeService->addSubmission($submission);
+
+        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
+            AuthorizationService::READ_SUBMISSION_ACTION));
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
+            AuthorizationService::UPDATE_SUBMISSION_ACTION));
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
+            AuthorizationService::MANAGE_ACTION));
+
+        $this->formalizeService->removeSubmission($submission);
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
+            AuthorizationService::READ_SUBMISSION_ACTION));
+    }
+
+    public function testSubmissionLevelAuthorizationAddAndUpdateSubmissionResource()
+    {
+        $form = $this->testEntityManager->addForm(
+            grantBasedSubmissionAuthorization: true,
+            allowedSubmissionStates: Submission::SUBMISSION_STATE_DRAFT | Submission::SUBMISSION_STATE_SUBMITTED,
+            actionsAllowedWhenSubmitted: [AuthorizationService::READ_SUBMISSION_ACTION],
+        );
+
+        // for draft submissions, the creator should have a manage grant
+        $submission = new Submission();
+        $submission->setForm($form);
+        $submission->setSubmissionState(Submission::SUBMISSION_STATE_DRAFT);
 
         $submission = $this->formalizeService->addSubmission($submission);
 
@@ -49,8 +87,9 @@ class AuthorizationIntegrationTest extends AbstractTestCase
             ResourceActionGrantService::MANAGE_ACTION));
         $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
-            AuthorizationService::UPDATE_SUBMISSION_ACTION));
+            AuthorizationService::UPDATE_SUBMISSION_ACTION)); // manage implies update
 
+        $this->authorizationService->clearCaches();
         $this->formalizeService->removeSubmission($submission);
         $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
@@ -58,12 +97,40 @@ class AuthorizationIntegrationTest extends AbstractTestCase
         $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
             AuthorizationService::UPDATE_SUBMISSION_ACTION));
+
+        // for submitted submissions, the creator should have only grants for the form's actionsAllowedWhenSubmitted.
+        // test transitioning from draft to submitted state on update (PATCH):
+        $this->authorizationService->clearCaches();
+        $submission = $this->formalizeService->addSubmission($submission);
+        $previousSubmission = clone $submission;
+
+        $submission->setDataFeedElement('{"givenName": "John", "familyName": "Doe"}');
+        $submission->setSubmissionState(Submission::SUBMISSION_STATE_SUBMITTED);
+
+        $this->authorizationService->clearCaches();
+        $submission = $this->formalizeService->updateSubmission($submission, $previousSubmission);
+
+        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
+            AuthorizationService::READ_SUBMISSION_ACTION));
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
+            AuthorizationService::UPDATE_SUBMISSION_ACTION));
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
+            AuthorizationService::MANAGE_ACTION));
+
+        $this->authorizationService->clearCaches();
+        $this->formalizeService->removeSubmission($submission);
+        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
+            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission->getIdentifier(),
+            AuthorizationService::READ_SUBMISSION_ACTION));
     }
 
     public function testNoSubmissionLevelAuthorizationAddAndRemoveSubmissionResource()
     {
         // No grants at all for creator-based submission authorization
-        $form = $this->addFormWithSubmissionLevelAuthorization(false);
+        $form = $this->testEntityManager->addForm(grantBasedSubmissionAuthorization: false);
 
         $submission = new Submission();
         $submission->setDataFeedElement('{"givenName": "John", "familyName": "Doe"}');
@@ -78,7 +145,10 @@ class AuthorizationIntegrationTest extends AbstractTestCase
 
     public function testSubmissionLevelAuthorizationRemoveAllSubmissionResourcesOnRequest(): void
     {
-        $form = $this->addFormWithSubmissionLevelAuthorization();
+        $form = $this->testEntityManager->addForm(
+            grantBasedSubmissionAuthorization: true,
+            actionsAllowedWhenSubmitted: [AuthorizationService::READ_SUBMISSION_ACTION],
+        );
 
         $submission1 = new Submission();
         $submission1->setDataFeedElement('{"givenName": "John", "familyName": "Doe"}');
@@ -99,19 +169,10 @@ class AuthorizationIntegrationTest extends AbstractTestCase
 
         $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(),
-            ResourceActionGrantService::MANAGE_ACTION));
-        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(),
             AuthorizationService::READ_SUBMISSION_ACTION));
         $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(),
-            ResourceActionGrantService::MANAGE_ACTION));
-        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(),
             AuthorizationService::READ_SUBMISSION_ACTION));
-        $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(),
-            ResourceActionGrantService::MANAGE_ACTION));
         $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(),
             AuthorizationService::READ_SUBMISSION_ACTION));
@@ -120,19 +181,10 @@ class AuthorizationIntegrationTest extends AbstractTestCase
 
         $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(),
-            ResourceActionGrantService::MANAGE_ACTION));
-        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(),
             AuthorizationService::READ_SUBMISSION_ACTION));
         $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(),
-            ResourceActionGrantService::MANAGE_ACTION));
-        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(),
             AuthorizationService::READ_SUBMISSION_ACTION));
-        $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
-            AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(),
-            ResourceActionGrantService::MANAGE_ACTION));
         $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(),
             AuthorizationService::READ_SUBMISSION_ACTION));
@@ -144,11 +196,10 @@ class AuthorizationIntegrationTest extends AbstractTestCase
 
     public function testSubmissionLevelAuthorizationRemoveAllSubmissionResourcesOnDeleteForm()
     {
-        $form = new Form();
-        $form->setName('Testform');
-        $form->setGrantBasedSubmissionAuthorization(true);
-
-        $form = $this->formalizeService->addForm($form);
+        $form = $this->testEntityManager->addForm(
+            grantBasedSubmissionAuthorization: true,
+            actionsAllowedWhenSubmitted: [AuthorizationService::READ_SUBMISSION_ACTION],
+        );
 
         $submission1 = new Submission();
         $submission1->setDataFeedElement('{"givenName": "John", "familyName": "Doe"}');
@@ -165,29 +216,24 @@ class AuthorizationIntegrationTest extends AbstractTestCase
 
         $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(),
-            ResourceActionGrantService::MANAGE_ACTION));
+            AuthorizationService::READ_SUBMISSION_ACTION));
         $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(),
-            ResourceActionGrantService::MANAGE_ACTION));
+            AuthorizationService::READ_SUBMISSION_ACTION));
         $this->assertTrue($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(),
-            ResourceActionGrantService::MANAGE_ACTION));
+            AuthorizationService::READ_SUBMISSION_ACTION));
 
         $this->formalizeService->removeForm($form);
 
         $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission1->getIdentifier(),
-            ResourceActionGrantService::MANAGE_ACTION));
+            AuthorizationService::READ_SUBMISSION_ACTION));
         $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission2->getIdentifier(),
-            ResourceActionGrantService::MANAGE_ACTION));
+            AuthorizationService::READ_SUBMISSION_ACTION));
         $this->assertFalse($this->resourceActionGrantService->isCurrentUserGrantedItemAction(
             AuthorizationService::SUBMISSION_RESOURCE_CLASS, $submission3->getIdentifier(),
-            ResourceActionGrantService::MANAGE_ACTION));
-    }
-
-    private function addFormWithSubmissionLevelAuthorization(bool $grantBasedSubmissionAuthorization = true): Form
-    {
-        return $this->testEntityManager->addForm('Testform', null, $grantBasedSubmissionAuthorization);
+            AuthorizationService::READ_SUBMISSION_ACTION));
     }
 }